% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_format.R
\name{in_format}
\alias{in_format}
\title{Calls rmarkdown::render with passed in output format code}
\usage{
in_format(x)
}
\arguments{
\item{x}{A numeric, non-decimal value to represent output format: 3 for "pdf_document", 33 for "html_document", and any other integer for "word_document".}
}
\value{
The rmarkdown::render function called with the interpreted output format from x
}
\description{
Calls rmarkdown::render with passed in output format code
}
\examples{
x <- 3
in_format(3)

x <- 33
in_format(x)
}
